# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "Criar novas dailyes, listar e fazer perguntas no HelpCenter"
  version: "1.0.0"
  title: "Sem Contrato Api"
  contact:
    name: "Suporte"
    email: "felipe.santos@compasso.com.br"
host: "reborn100contrato.azurewebsites.net"
tags:
- name: "Auth"
- name: "User"
- name: "Configuration"
- name: "Daily"
- name: "HelpCenter"
- name: "HelpCenterASK"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /users/authenticate:
    post:
      tags:
      - "Auth"
      summary: "Autenticação"
      description: "Autenticação de Usuario"
      operationId: "AuthUser"
      parameters:
      - in: "body"
        name: "docs"
        required: false
        schema:
          $ref: "#/definitions/Auth"
      responses:
        200:
          description: "Cadastro efetuado com sucesso"
          schema:
            $ref: "#/definitions/Authenticate"
        400:
          description: "Email ou senha inválidos"
        500:
          description: "Erro interno no servidor"
      x-swagger-router-controller: "Auth"
  /admin/export/users:
    get:
      tags:
      - "User"
      summary: "Lista todos Usuarios"
      description: "Lista todos os Usuario cadastrado do BD"
      operationId: "UserExport"
      parameters:
      - name: "authorization"
        in: "header"
        required: true
        type: "string"
      responses:
        200:
          description: "Cadastro efetuado com sucesso"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Authenticate"
        400:
          description: "Email ou senha inválidos"
        500:
          description: "Erro interno no servidor"
      security:
      - Bearer: []
      x-swagger-router-controller: "User"
  /dailys:
    get:
      tags:
      - "Daily"
      summary: "Lista todas Dailyes"
      description: "Lista as ultimas 10 daily cadastradas"
      operationId: "listaTodasDaily"
      parameters:
      - name: "authorization"
        in: "header"
        required: true
        type: "string"
      responses:
        201:
          description: "Cadastro efetuado com sucesso"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Daily"
        401:
          description: "Token Inválido"
        500:
          description: "Erro interno no servidor"
      security:
      - Bearer: []
      x-swagger-router-controller: "Daily"
  /dailys/{id}:
    get:
      tags:
      - "Daily"
      summary: "Procura por ID Daily"
      description: "Procura por ID Daily"
      operationId: "sendDaily"
      parameters:
      - name: "authorization"
        in: "header"
        required: true
        type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Cadastro efetuado com sucesso"
          schema:
            $ref: "#/definitions/Daily"
        400:
          description: "Houve Algum problema na hora de mostrar os dados da daily\
            \ favor olhar o log"
        401:
          description: "Token Inválido"
        500:
          description: "Erro interno no servidor"
      security:
      - Bearer: []
      x-swagger-router-controller: "Daily"
  /dailys/user/{id}:
    get:
      tags:
      - "Daily"
      summary: "Procura por ID do Usuerio"
      description: "Verifica se a daily já foi cadastrada ou não"
      operationId: "sendUserDaily"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: ""
        400:
          description: "Você já cadastrou sua daily"
        401:
          description: "Token Inválido"
        500:
          description: "Erro interno no servidor"
      x-swagger-router-controller: "Daily"
  /dailys/list/{date}/{page}:
    get:
      tags:
      - "Daily"
      summary: "Lista Dailyes do dia"
      description: "Lista as dailys do dia"
      operationId: "listaDailyDia"
      parameters:
      - name: "date"
        in: "path"
        required: true
        type: "string"
        format: "date"
      - name: "page"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "docs"
        required: false
        schema:
          $ref: "#/definitions/Daily"
      responses:
        200:
          description: "Lista daily do dia"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Daily"
        500:
          description: "Erro interno no servidor"
      security:
      - Bearer: []
      x-swagger-router-controller: "Daily"
  /dailys/daily/:
    post:
      tags:
      - "Daily"
      summary: "Cadastrar Daily"
      description: "Cadastrar Daily do dia"
      operationId: "cadastrarDailyDia"
      parameters:
      - name: "authorization"
        in: "header"
        required: true
        type: "string"
      - in: "body"
        name: "docs"
        required: true
        schema:
          $ref: "#/definitions/DailyCadastro"
      responses:
        200:
          description: "Lista daily do dia"
          schema:
            $ref: "#/definitions/Daily"
        400:
          description: "USUARIO não existente"
        500:
          description: "Erro interno no servidor"
      x-swagger-router-controller: "Daily"
  /dailys/daily/{id}:
    put:
      tags:
      - "Daily"
      summary: "Atualiza Daily"
      description: "Atualizar Daily"
      operationId: "atualizarDailyDia"
      parameters:
      - name: "authorization"
        in: "header"
        required: true
        type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "docs"
        required: true
        schema:
          $ref: "#/definitions/DailyEditar"
      responses:
        200:
          description: "Lista daily do dia"
          schema:
            $ref: "#/definitions/Daily"
        400:
          description: "USUARIO não existente"
        500:
          description: "Erro interno no servidor"
      security:
      - Bearer: []
      x-swagger-router-controller: "Daily"
  /admin/export/daylies:
    get:
      tags:
      - "Daily"
      summary: "Lista Todas DAYLIES do Banco"
      description: "Lista Todas DAYLIES do Banco precisa ser Admin"
      operationId: "AdminListaDailyDia"
      parameters:
      - name: "authorization"
        in: "header"
        required: true
        type: "string"
      - in: "body"
        name: "docs"
        required: false
        schema:
          $ref: "#/definitions/Daily"
      responses:
        200:
          description: "Lista todas Daily"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Daily"
        500:
          description: "Erro interno no servidor"
      security:
      - Bearer: []
      x-swagger-router-controller: "Daily"
  /helps/list/post/{page}:
    get:
      tags:
      - "HelpCenter"
      summary: "Lista HelpCenter"
      description: "Lista HelpCenter Paginada"
      operationId: "ListaHelpCenter"
      parameters:
      - name: "page"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Lista daily do dia"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/HelpCenterList"
        400:
          description: "Houve Algum problema na hora de listar o usuario favor olhar\
            \ o log"
        500:
          description: "Erro interno no servidor"
      x-swagger-router-controller: "HelpCenter"
  /helps/list/{id}:
    get:
      tags:
      - "HelpCenter"
      summary: "Lista HelpCenter"
      description: "Procura HelpCenter"
      operationId: "sendHelpCenter"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Lista daily do dia"
          schema:
            $ref: "#/definitions/HelpCenterAddResp"
        400:
          description: "Houve Algum problema na hora de listar o usuario favor olhar\
            \ o log"
        500:
          description: "Erro interno no servidor"
      x-swagger-router-controller: "HelpCenter"
  /helps/post/:
    post:
      tags:
      - "HelpCenter"
      summary: "Cadastra HelpCenter"
      description: "Cadastra HelpCenter Paginada"
      operationId: "CadastraHelpCenter"
      parameters:
      - name: "authorization"
        in: "header"
        required: true
        type: "string"
      - in: "body"
        name: "docs"
        required: true
        schema:
          $ref: "#/definitions/HelpCenterAdd"
      responses:
        200:
          description: "Lista daily do dia"
          schema:
            $ref: "#/definitions/HelpCenterAddResp"
        400:
          description: "Houve Algum problema na hora de listar o usuario favor olhar\
            \ o log"
        500:
          description: "Erro interno no servidor"
      security:
      - Bearer: []
      x-swagger-router-controller: "HelpCenter"
  /helps/post/{id}:
    delete:
      tags:
      - "HelpCenter"
      summary: "Deletar HelpCenter"
      description: "Deletar HelpCenter"
      operationId: "DeleteHelpCenter"
      parameters:
      - name: "authorization"
        in: "header"
        required: true
        type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Deletado"
          schema:
            $ref: "#/definitions/HelpCenterAddResp"
        401:
          description: "Token Inválido"
        500:
          description: "Não é ADMIN"
      security:
      - Bearer: []
      x-swagger-router-controller: "HelpCenter"
  /helps/list/ask/{page}:
    get:
      tags:
      - "HelpCenterASK"
      summary: "Lista HelpCenterASK"
      description: "Lista HelpCenterASK"
      operationId: "ListHelpCenterAsk"
      parameters:
      - name: "page"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Lista HelpCenter de pergunta"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/HelpCenterASKList"
        400:
          description: "Houve Algum problema na hora de listar o usuario favor olhar\
            \ o log"
        500:
          description: "Erro interno no servidor"
      x-swagger-router-controller: "HelpCenterASK"
  /helps/list/answers/{idQuestion}/{page}:
    get:
      tags:
      - "HelpCenterASK"
      summary: "Procura HelpCenterASK"
      description: "Procura HelpCenterASK (Resposta)"
      operationId: "sendHelpCenterASKResposta"
      parameters:
      - name: "idQuestion"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - name: "page"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Lista todas respostas da pergunta"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/HelpCenterASKSend"
        400:
          description: "Houve Algum problema na hora de listar o usuario favor olhar\
            \ o log"
        500:
          description: "Erro interno no servidor"
      x-swagger-router-controller: "HelpCenterASK"
  /admin/export/helpCenter:
    get:
      tags:
      - "HelpCenter"
      summary: "Lista Todas HelpCenter"
      description: "Lista Todas HelpCenter precisa ser admin"
      operationId: "AdminListaHelpCenter"
      parameters:
      - name: "authorization"
        in: "header"
        required: true
        type: "string"
      responses:
        200:
          description: "Lista Todas HelpCenter do Banco"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/HelpCenterAddResp"
        400:
          description: "Houve Algum problema na hora de listar o usuario favor olhar\
            \ o log"
        500:
          description: "Erro interno no servidor"
      security:
      - Bearer: []
      x-swagger-router-controller: "HelpCenter"
  /helps/ask/{id}:
    get:
      tags:
      - "HelpCenterASK"
      summary: "Procura HelpCenterASK"
      description: "Procura HelpCenterASK (Resposta)"
      operationId: "sendHelpCenterASK"
      parameters:
      - name: "authorization"
        in: "header"
        required: true
        type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Lista HelpCenterASK"
          schema:
            $ref: "#/definitions/HelpCenterASKSend"
        400:
          description: "Houve Algum problema na hora de listar o usuario favor olhar\
            \ o log"
        500:
          description: "Erro interno no servidor"
      security:
      - Bearer: []
      x-swagger-router-controller: "HelpCenterASK"
  /admin/export/helpCenterAnswer:
    get:
      tags:
      - "HelpCenterASK"
      summary: "Lista Todas HelpCenterASK"
      description: "Lista Todas HelpCenterASK precisa ser admin"
      operationId: "AdminListaHelpCenterASK"
      parameters:
      - name: "authorization"
        in: "header"
        required: true
        type: "string"
      responses:
        200:
          description: "Lista Todas HelpCenter do Banco"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/HelpCenterASKList"
        400:
          description: "Houve Algum problema na hora de listar o usuario favor olhar\
            \ o log"
        500:
          description: "Erro interno no servidor"
      security:
      - Bearer: []
      x-swagger-router-controller: "HelpCenterASK"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  Auth:
    type: "object"
    properties:
      email:
        type: "string"
        example: "batata@gmail.com"
      password:
        type: "string"
        example: "123123"
    example:
      password: "123123"
      email: "batata@gmail.com"
  Authenticate:
    type: "object"
    properties:
      isAdmin:
        type: "boolean"
        example: false
      _id:
        type: "string"
        example: "5d33fcb3fa5500198847300d"
      name:
        type: "string"
        example: "João P"
      lastName:
        type: "string"
        example: "batata"
      userName:
        type: "string"
        example: "batata"
      email:
        type: "string"
        example: "batata@gmail.com"
      dateOfBirth:
        type: "string"
        format: "date-time"
      password:
        type: "string"
        example: "a6361743ee6a60423a1879014f38fe37d363c840adc9e37deb5609c32fc8f2ea"
      createdAt:
        type: "string"
        format: "date-time"
    example:
      lastName: "batata"
      createdAt: "2000-01-23T04:56:07.000+00:00"
      password: "a6361743ee6a60423a1879014f38fe37d363c840adc9e37deb5609c32fc8f2ea"
      name: "João P"
      dateOfBirth: "2000-01-23T04:56:07.000+00:00"
      isAdmin: false
      _id: "5d33fcb3fa5500198847300d"
      userName: "batata"
      email: "batata@gmail.com"
  Daily:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "5d41ca343c941f3bf0fddcf2"
      id_user:
        type: "string"
        example: "5d41c9b63c941f3bf0fddcf1"
      yesterday:
        type: "string"
        example: "Testes e correções de bugs."
      today:
        type: "string"
        example: "Testes e desenvolvimento."
      impediment:
        type: "string"
        example: "Nenhum impedimento."
      date:
        type: "string"
        format: "date-time"
    example:
      yesterday: "Testes e correções de bugs."
      date: "2000-01-23T04:56:07.000+00:00"
      today: "Testes e desenvolvimento."
      impediment: "Nenhum impedimento."
      _id: "5d41ca343c941f3bf0fddcf2"
      id_user: "5d41c9b63c941f3bf0fddcf1"
  DailyCadastro:
    type: "object"
    properties:
      email:
        type: "string"
        example: "teste@gmail.com"
      yesterday:
        type: "string"
        example: "Testes e correções de bugs."
      today:
        type: "string"
        example: "Testes e desenvolvimento."
      impediment:
        type: "string"
        example: "Nenhum impedimento."
    example:
      yesterday: "Testes e correções de bugs."
      today: "Testes e desenvolvimento."
      impediment: "Nenhum impedimento."
      email: "teste@gmail.com"
  DailyEditar:
    type: "object"
    properties:
      yesterday:
        type: "string"
        example: "Testes e correções de bugs."
      today:
        type: "string"
        example: "Testes e desenvolvimento."
      impediment:
        type: "string"
        example: "Nenhum impedimento."
    example:
      yesterday: "Testes e correções de bugs."
      today: "Testes e desenvolvimento."
      impediment: "Nenhum impedimento."
  HelpCenterList:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "5d30b2610ae9530036eec2ae"
      title:
        type: "string"
        example: "Deploy"
      desc:
        type: "string"
        example: "Deploy do azure não funciona bem da erro no compilador, no express"
      date:
        type: "string"
        format: "date-time"
      id_user:
        type: "string"
        example: "5d30b1110ae9530036eec2a9"
      owner:
        type: "string"
        example: "Felipe Santos"
    example:
      date: "2000-01-23T04:56:07.000+00:00"
      owner: "Felipe Santos"
      _id: "5d30b2610ae9530036eec2ae"
      id_user: "5d30b1110ae9530036eec2a9"
      title: "Deploy"
      desc: "Deploy do azure não funciona bem da erro no compilador, no express"
  HelpCenterAdd:
    type: "object"
    properties:
      id_user:
        type: "string"
        example: "5d30b2610ae9530036eec2ae"
      title:
        type: "string"
        example: "Deploy"
      desc:
        type: "string"
        example: "Deploy do azure não funciona bem da erro no compilador, no express"
    example:
      id_user: "5d30b2610ae9530036eec2ae"
      title: "Deploy"
      desc: "Deploy do azure não funciona bem da erro no compilador, no express"
  HelpCenterAddResp:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "5d432d50a6d90d38c44061e6"
      id_user:
        type: "string"
        example: "5d30b2610ae9530036eec2ae"
      title:
        type: "string"
        example: "Deploy"
      desc:
        type: "string"
        example: "Deploy do azure não funciona bem da erro no compilador, no express"
      date:
        type: "string"
        format: "date-time"
    example:
      date: "2000-01-23T04:56:07.000+00:00"
      _id: "5d432d50a6d90d38c44061e6"
      id_user: "5d30b2610ae9530036eec2ae"
      title: "Deploy"
      desc: "Deploy do azure não funciona bem da erro no compilador, no express"
  HelpCenterASKList:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "5d30b35a0ae9530036eec2b1"
      desc:
        type: "string"
        example: "Não sei."
      date:
        type: "string"
        format: "date-time"
      id_user:
        type: "string"
        example: "5d30b01b0ae9530036eec2a4"
      id_helpCenter:
        type: "string"
        example: "5d30b2610ae9530036eec2ae"
      help:
        type: "string"
        example: "help"
      owner:
        type: "string"
        example: "Willian Froes"
    example:
      date: "2000-01-23T04:56:07.000+00:00"
      owner: "Willian Froes"
      help: "help"
      _id: "5d30b35a0ae9530036eec2b1"
      id_user: "5d30b01b0ae9530036eec2a4"
      id_helpCenter: "5d30b2610ae9530036eec2ae"
      desc: "Não sei."
  HelpCenterASKSend:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "5d30b35a0ae9530036eec2b1"
      id_user:
        type: "string"
        example: "5d30b01b0ae9530036eec2a4"
      id_helpCenter:
        type: "string"
        example: "5d30b2610ae9530036eec2ae"
      desc:
        type: "string"
        example: "Não sei."
      date:
        type: "string"
        format: "date-time"
    example:
      date: "2000-01-23T04:56:07.000+00:00"
      _id: "5d30b35a0ae9530036eec2b1"
      id_user: "5d30b01b0ae9530036eec2a4"
      id_helpCenter: "5d30b2610ae9530036eec2ae"
      desc: "Não sei."